+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc -b && vite build",
-    "lint": "eslint \"src/**/*.{ts,tsx}\" --fix",
-    "format": "prettier --write \"src/**/*.{ts,tsx,js,jsx,json,css,md}\"",
+    "lint": "eslint \"src/**\" --fix",
+    "format": "prettier --write \"src/**\"",
     "type-check": "tsc --noEmit",
     "preview": "vite preview"
   },
@@ -38,10 +38,11 @@
     "eslint-plugin-react": "^7.34.3",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.7",
+    "gh-pages": "^6.1.1",
     "postcss": "^8.4.38",
     "prettier": "^3.3.2",
     "sass": "^1.77.6",
-    "typescript": "^4.7.4",
+    "typescript": "^5.5.2",
     "vite": "^5.3.1"
   }
 }
