+++ src/App/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import App from "./App";
 import "./index.scss";
 import { Provider } from "react-redux";
-import store from "../store";
+import store from "../store/store";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <Provider store={store}>
