+++ vite.config.ts
@@ -1,6 +1,7 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react-swc";
 import autoprefixer from "autoprefixer";
+import { fileURLToPath, URL } from 'url';
 
 export default defineConfig({
   plugins: [react()],
@@ -14,7 +15,8 @@ export default defineConfig({
   },
   resolve: {
     alias: {
-      "@": "/src",
-    },
+      '@': fileURLToPath(new URL('./src/', import.meta.url)),
+    }
   },
+
 });
